Add present and futur tile map
