+++ Assets/Palettes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5d6c1edee9451f4cab251fe5927e24d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
