+++ Assets/Palettes/simple_bridge.asset.meta
