+++ Assets/Palettes/simple_dirtblock.asset
