+++ Assets/Palettes/simple_grassblock.asset
