+++ Assets/Palettes/simple_grassblock.png.meta
